reorganize tests: semantic analysis
