+++ LUIECompilerTests/SemanticAnalysis/DefinitionTest.cs
@@ -1,0 +1,56 @@
+using Antlr4.Runtime;
+using Antlr4.Runtime.Tree;
+using LUIECompiler.Common;
+using LUIECompiler.Common.Errors;
+using LUIECompiler.Common.Symbols;
+using LUIECompiler.SemanticAnalysis;
+
+namespace LUIECompilerTests.SemanticAnalysis;
+
+[TestClass]
+public class DefinitionTest
+{
+    public const string InputSimple =
+        "qubit c;\n" +
+        "qubit b;\n" +
+        "x a;\n" +
+        "qubit c;\n" +
+        "qif a do\n" +
+        "skip;\n" +
+        "end";
+
+    /// <summary>
+    /// Test that already defined identifiers are correctly reported.
+    /// </summary>
+    [TestMethod]
+    public void RedefineErrorTest()
+    {
+        var walker = Utils.GetWalker();
+        var parser = Utils.GetParser(InputSimple);
+        var analysis = new DeclarationAnalysisListener();
+        walker.Walk(analysis, parser.parse());
+        var error = analysis.Error;
+
+        Assert.IsTrue(error.ContainsCriticalError);
+
+        Assert.IsTrue(error.Errors.Any(e => e is RedefineError && e.Line == 4));
+    }
+
+    /// <summary>
+    /// Tests that undefined identifiers are correctly reported.
+    /// </summary>
+    [TestMethod]
+    public void UndefinedErrorTest()
+    {
+        var walker = Utils.GetWalker();
+        var parser = Utils.GetParser(InputSimple);
+        var analysis = new DeclarationAnalysisListener();
+        walker.Walk(analysis, parser.parse());
+        var error = analysis.Error;
+
+        Assert.IsTrue(error.ContainsCriticalError);
+
+        Assert.IsTrue(error.Errors.Any(e => e is UndefinedError && e.Line == 3));
+        Assert.IsTrue(error.Errors.Any(e => e is UndefinedError && e.Line == 5));
+    }
+}
